refactor!(mediums): update root `artworkMediumTypes` field to `artworkMediums`
