+++ src/lib/artworkMediums.ts
