+++ src/schema/v2/schema.ts
@@ -102,7 +102,7 @@ import { DeleteArtworkImageMutation } from "./deleteArtworkImageMutation"
 
 import { ArtworkOrEditionSetType } from "schema/v2/artworkOrEditionSet"
 import { AuctionResult } from "./auction_result"
-import ArtworkMediumTypes from "./artworkMediumTypes"
+import ArtworkMediums from "./artworkMediums"
 
 const PrincipalFieldDirective = new GraphQLDirective({
   name: "principalField",
@@ -111,7 +111,7 @@ const PrincipalFieldDirective = new GraphQLDirective({
 
 const rootFields = {
   artworkAttributionClasses: ArtworkAttributionClasses,
-  artworkMediumTypes: ArtworkMediumTypes,
+  artworkMediums: ArtworkMediums,
   auctionResult: AuctionResult,
   article: Article,
   articles: Articles,
